+++ CHANGELOG.md
@@ -45,6 +45,7 @@ All notable changes to this project will be documented in this file.
 - *(pdf)* Verifier about ai injection working
 - *(pdf)* Generating question from secure pdf created
 - *(id)* Added a pattern for identification of the cards
+- *(ai)* Improved to clean useless char data for input
 
 ### ⚙️ Miscellaneous Tasks